[Refactor/#12] refactor package structure
