--- src/main/java/com/rockoon/domain/board/chatting/service/ChattingCommandService.java
@@ -1,4 +1,0 @@
-package com.rockoon.domain.board.chatting.service;
-
-public interface ChattingCommandService {
-}
