--- src/main/java/com/rockoon/domain/board/chatting/service/ChattingQueryServiceImpl.java
@@ -1,4 +1,0 @@
-package com.rockoon.domain.board.chatting.service;
-
-public class ChattingQueryServiceImpl {
-}
