+++ src/main/java/com/rockoon/domain/board/entity/Board.java
@@ -1,7 +1,7 @@
 package com.rockoon.domain.board.entity;
 
 import com.rockoon.domain.member.entity.Member;
-import com.rockoon.global.entity.BaseTimeEntity;
+import com.rockoon.domain.auditing.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
