+++ src/main/java/com/rockoon/domain/board/service/promotion/PromotionQueryServiceImpl.java
@@ -1,7 +1,7 @@
-package com.rockoon.domain.board.promotion.service;
+package com.rockoon.domain.board.service.promotion;
 
-import com.rockoon.domain.board.promotion.entity.Promotion;
-import com.rockoon.domain.board.promotion.repository.PromotionRepository;
+import com.rockoon.domain.board.entity.Promotion;
+import com.rockoon.domain.board.repository.PromotionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
