+++ src/main/java/com/rockoon/domain/image/entity/Image.java
@@ -1,7 +1,7 @@
 package com.rockoon.domain.image.entity;
 
 import com.rockoon.domain.board.entity.Board;
-import com.rockoon.global.entity.BaseTimeEntity;
+import com.rockoon.domain.auditing.entity.BaseTimeEntity;
 import com.rockoon.domain.image.dto.ImageRequest;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
