+++ src/main/java/com/rockoon/domain/option/entity/Option.java
@@ -1,7 +1,7 @@
 package com.rockoon.domain.option.entity;
 
 import com.rockoon.domain.board.entity.Board;
-import com.rockoon.global.entity.BaseTimeEntity;
+import com.rockoon.domain.auditing.entity.BaseTimeEntity;
 import com.rockoon.domain.option.dto.OptionRequest;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
