+++ src/main/java/com/rockoon/domain/team/entity/TeamMember.java
@@ -1,7 +1,7 @@
 package com.rockoon.domain.team.entity;
 
 import com.rockoon.domain.member.entity.Member;
-import com.rockoon.global.entity.BaseTimeEntity;
+import com.rockoon.domain.auditing.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
