+++ src/test/java/com/rockoon/domain/member/repository/MemberRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.rockoon.domain.member.repository;
 
 import com.rockoon.domain.member.entity.Member;
-import com.rockoon.global.test.DatabaseCleanUp;
+import com.rockoon.global.config.test.DatabaseCleanUp;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
