+++ src/test/java/com/rockoon/domain/member/service/MemberCommandServiceTest.java
@@ -4,7 +4,7 @@ import com.rockoon.domain.member.dto.MemberRequest.MemberModifyDto;
 import com.rockoon.domain.member.dto.MemberRequest.MemberRegisterDto;
 import com.rockoon.domain.member.entity.Member;
 import com.rockoon.domain.member.repository.MemberRepository;
-import com.rockoon.global.test.DatabaseCleanUp;
+import com.rockoon.global.config.test.DatabaseCleanUp;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
