+++ src/test/java/com/rockoon/domain/team/service/TeamCommandServiceTest.java
@@ -7,7 +7,7 @@ import com.rockoon.domain.team.entity.Team;
 import com.rockoon.domain.team.entity.TeamMember;
 import com.rockoon.domain.team.repository.TeamMemberRepository;
 import com.rockoon.domain.team.repository.TeamRepository;
-import com.rockoon.global.test.DatabaseCleanUp;
+import com.rockoon.global.config.test.DatabaseCleanUp;
 import com.rockoon.domain.team.dto.TeamRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterEach;
